requirements.txt
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.0.3
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-datastore==1.8.0
googleapis-common-protos==1.6.0
grpcio==1.21.1
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
MarkupSafe==1.1.1
more-itertools==7.0.0
packaging==19.0
passlib==1.7.1
pipenv==2018.11.26
pluggy==0.12.0
protobuf==3.8.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.0
pytest==4.6.3
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
